refactor into single error dialog
