--- app/src/main/java/com/manichord/mgit/dialogs/ExceptionDialog.kt
@@ -1,75 +1,0 @@
-package com.manichord.mgit.dialogs
-
-import android.app.AlertDialog
-import android.app.Dialog
-import android.os.Bundle
-import android.support.annotation.StringRes
-import android.widget.Button
-import io.sentry.Sentry
-import kotlinx.android.synthetic.main.dialog_exception.view.*
-import me.sheimi.android.views.SheimiDialogFragment
-import me.sheimi.sgit.BuildConfig
-import me.sheimi.sgit.R
-import me.sheimi.sgit.dialogs.DummyDialogListener
-import timber.log.Timber
-import kotlin.Exception
-
-class ExceptionDialog : SheimiDialogFragment() {
-    private var mThrowable: Throwable? = null
-    @StringRes
-    private var mErrorRes: Int = 0
-    @StringRes
-    var errorTitleRes: Int = 0
-
-    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
-        super.onCreateDialog(savedInstanceState)
-
-        val builder = AlertDialog.Builder(rawActivity)
-        val inflater = rawActivity.layoutInflater
-        val layout = inflater.inflate(R.layout.dialog_exception, null)
-        val details = when (mThrowable) {
-            is Exception -> { (mThrowable as Exception).message }
-            else -> ""
-        }
-        layout.error_message.setText(getString(mErrorRes) + "\n" + details)
-
-        builder.setView(layout)
-
-        // set button listener
-        builder.setTitle(if (errorTitleRes != 0) errorTitleRes else R.string.dialog_error_title)
-        builder.setNegativeButton(getString(R.string.label_cancel),
-                DummyDialogListener())
-        builder.setPositiveButton(
-                getString(R.string.dialog_error_send_report),
-                DummyDialogListener())
-
-        return builder.create()
-    }
-
-    override fun onStart() {
-        super.onStart()
-        val dialog = dialog as AlertDialog
-        val positiveButton = dialog.getButton(Dialog.BUTTON_POSITIVE) as Button
-        positiveButton.setOnClickListener {
-            if (BuildConfig.DEBUG) {
-                // when debugging just log the exception
-                Timber.e(mThrowable);
-            } else {
-                Sentry.capture(mThrowable);
-            }
-            dismiss()
-        }
-        val negativeButton = dialog.getButton(Dialog.BUTTON_NEGATIVE)
-        negativeButton.setOnClickListener {
-            dismiss()
-        }
-    }
-
-    fun setThrowable(throwable: Throwable?) {
-        mThrowable = throwable
-    }
-
-    fun setErrorRes(@StringRes errorRes: Int) {
-        mErrorRes = errorRes
-    }
-}
