+++ app/src/main/java/me/sheimi/android/utils/BasicFunctions.java
@@ -5,7 +5,6 @@ import android.support.annotation.StringRes;
 import android.widget.ImageView;
 
 import com.manichord.mgit.dialogs.ErrorDialog;
-import com.manichord.mgit.dialogs.ExceptionDialog;
 import com.nostra13.universalimageloader.core.ImageLoader;
 
 import org.jetbrains.annotations.NotNull;
@@ -69,18 +68,18 @@ public class BasicFunctions {
     }
 
     public static void showError(@NonNull @NotNull SheimiFragmentActivity activity, @StringRes final int errorTitleRes, @StringRes final int errorRes) {
-        ErrorDialog exceptionDialog = new ErrorDialog();
-        exceptionDialog.setErrorRes(errorRes);
-        exceptionDialog.setErrorTitleRes(errorTitleRes);
-        exceptionDialog.show(activity.getSupportFragmentManager(), "error-dialog");
+        ErrorDialog errorDialog = new ErrorDialog();
+        errorDialog.setErrorRes(errorRes);
+        errorDialog.setErrorTitleRes(errorTitleRes);
+        errorDialog.show(activity.getSupportFragmentManager(), "error-dialog");
     }
 
     public static void showException(@NonNull @NotNull SheimiFragmentActivity activity, Throwable throwable, @StringRes final int errorTitleRes, @StringRes final int errorRes) {
-        ExceptionDialog exceptionDialog = new ExceptionDialog();
-        exceptionDialog.setThrowable(throwable);
-        exceptionDialog.setErrorRes(errorRes);
-        exceptionDialog.setErrorTitleRes(errorTitleRes);
-        exceptionDialog.show(activity.getSupportFragmentManager(), "exception-dialog");
+        ErrorDialog errorDialog = new ErrorDialog();
+        errorDialog.setThrowable(throwable);
+        errorDialog.setErrorRes(errorRes);
+        errorDialog.setErrorTitleRes(errorTitleRes);
+        errorDialog.show(activity.getSupportFragmentManager(), "exception-dialog");
     }
