--- app/src/main/res/layout/dialog_exception.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-android:layout_width="match_parent"
-android:layout_height="match_parent"
-android:orientation="vertical">
-    <TextView
-        android:id="@+id/error_message"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:paddingTop="@dimen/activity_vertical_margin"
-        android:paddingStart="@dimen/large_horizontal_margin"
-        android:paddingLeft="@dimen/large_horizontal_margin"
-        android:text="@string/dialog_error_occurred_msg" />
-</LinearLayout>
